pkg: move `test/utils/nodes` pkg into `internal`
Trivial code movement again to trim down `test/utils` a bit more.
Only trivial changes and import path fixup.
