+++ test/e2e/serial/tests/configuration.go
@@ -45,6 +45,7 @@ import (
 	nropmcp "github.com/openshift-kni/numaresources-operator/pkg/machineconfigpools"
 	rteconfig "github.com/openshift-kni/numaresources-operator/rte/pkg/config"
 
+	"github.com/openshift-kni/numaresources-operator/internal/nodes"
 	e2ereslist "github.com/openshift-kni/numaresources-operator/internal/resourcelist"
 	"github.com/openshift-kni/numaresources-operator/internal/wait"
 
@@ -52,7 +53,6 @@ import (
 	"github.com/openshift-kni/numaresources-operator/test/utils/configuration"
 	e2efixture "github.com/openshift-kni/numaresources-operator/test/utils/fixture"
 	e2enrt "github.com/openshift-kni/numaresources-operator/test/utils/noderesourcetopologies"
-	e2enodes "github.com/openshift-kni/numaresources-operator/test/utils/nodes"
 	"github.com/openshift-kni/numaresources-operator/test/utils/nrosched"
 	"github.com/openshift-kni/numaresources-operator/test/utils/objects"
 	e2epadder "github.com/openshift-kni/numaresources-operator/test/utils/padder"
@@ -146,15 +146,15 @@ var _ = Describe("[serial][disruptive][slow] numaresources configuration managem
 			Expect(err).ToNot(HaveOccurred(), "cannot get %q in the cluster", objects.NROName())
 			initialNroOperObj := nroOperObj.DeepCopy()
 
-			workers, err := e2enodes.GetWorkerNodes(fxt.Client)
+			workers, err := nodes.GetWorkerNodes(fxt.Client)
 			Expect(err).ToNot(HaveOccurred())
 			// TODO choose randomly
 			targetedNode := workers[0]
 
-			unlabelFunc, err := labelNode(fxt.Client, e2enodes.GetLabelRoleMCPTest(), targetedNode.Name)
+			unlabelFunc, err := labelNode(fxt.Client, nodes.GetLabelRoleMCPTest(), targetedNode.Name)
 			Expect(err).ToNot(HaveOccurred())
 
-			labelFunc, err := unlabelNode(fxt.Client, e2enodes.GetLabelRoleWorker(), "", targetedNode.Name)
+			labelFunc, err := unlabelNode(fxt.Client, nodes.GetLabelRoleWorker(), "", targetedNode.Name)
 			Expect(err).ToNot(HaveOccurred())
 
 			defer func() {
@@ -213,18 +213,18 @@ var _ = Describe("[serial][disruptive][slow] numaresources configuration managem
 			mcp := objects.TestMCP()
 			By(fmt.Sprintf("creating new MCP: %q", mcp.Name))
 			// we must have this label in order to match other machine configs that are necessary for proper functionality
-			mcp.Labels = map[string]string{"machineconfiguration.openshift.io/role": e2enodes.RoleMCPTest}
+			mcp.Labels = map[string]string{"machineconfiguration.openshift.io/role": nodes.RoleMCPTest}
 			mcp.Spec.MachineConfigSelector = &metav1.LabelSelector{
 				MatchExpressions: []metav1.LabelSelectorRequirement{
 					{
 						Key:      "machineconfiguration.openshift.io/role",
 						Operator: metav1.LabelSelectorOpIn,
-						Values:   []string{e2enodes.RoleWorker, e2enodes.RoleMCPTest},
+						Values:   []string{nodes.RoleWorker, nodes.RoleMCPTest},
 					},
 				},
 			}
 			mcp.Spec.NodeSelector = &metav1.LabelSelector{
-				MatchLabels: map[string]string{e2enodes.GetLabelRoleMCPTest(): ""},
+				MatchLabels: map[string]string{nodes.GetLabelRoleMCPTest(): ""},
 			}
 
 			err = fxt.Client.Create(context.TODO(), mcp)
